move routes to file
