+++ resources/contribute-app/router/index.ts
@@ -1,0 +1,8 @@
+import { createRouter, createWebHistory } from 'vue-router';
+import routes from './routes';
+
+export default createRouter({
+    history: createWebHistory('/contribute'),
+    routes,
+});
+
